refactor(types): split dpp definitions
