+++ src/types/dpp/index.ts
@@ -1,0 +1,3 @@
+export * from './Lifecycle';
+export * from './Compliance';
+export * from './Product';
